Modify "my_strncat" function
The meaning of the "len" parameter in the my_strncat function is the size limit for copying characters from "from", not the size limit for "to" after copying.
Also, the "#define safestrcat(to, from) my_strncat(to, from, sizeof(to) - strlen(to) - 1)" has already imposed a limit on max based on the size of "to".
Modify the function to prevent truncation of content when too many bytes are passed to the my_strcat function.

Signed-off-by: songliang <YS.songliang@h3c.com>